+++ cfg.py
@@ -16,11 +16,3 @@ to_server_to_file = False
 
 # write loaded from server content to file
 from_server_to_file = False
-
-
-# -------------------------------------------------------------------------------
-# experimental features --- For release all must be = False
-# -------------------------------------------------------------------------------
-
-# use BIS custom nodes
-experimental_enable_bis_custom_nodes = False